# debugging perf impact of zero copy buf