master branch:
```
~/src/deno> wrk -d 30 http://127.0.0.1:4544/
Running 30s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   200.51us   52.46us   4.58ms   91.34%
    Req/Sec    23.61k     1.29k   26.01k    79.24%
  1414162 requests in 30.10s, 68.78MB read
Requests/sec:  46982.32
Transfer/sec:      2.29MB
```

with this patch:
```
~/src/deno> wrk -d 30 http://127.0.0.1:4544/
Running 30s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   181.70us   52.86us   4.81ms   92.58%
    Req/Sec    25.96k     1.30k   28.17k    84.72%
  1554967 requests in 30.10s, 75.63MB read
Requests/sec:  51660.57
Transfer/sec:      2.51MB
```